+++ src/main.cpp
@@ -9,32 +9,8 @@
  *  Processes the security system status and implement the Homie convention.
  *
  *  See README.md
- *
- *  Wiring:
- *      DSC Aux(+) ---+--- esp8266 NodeMCU Vin pin
- *                    |
- *                    +--- 5v voltage regulator --- esp8266 Wemos D1 Mini 5v pin
- *
- *      DSC Aux(-) --- esp8266 Ground
- *
- *                                         +--- dscClockPin (esp8266: D1, D2, D8)
- *      DSC Yellow --- 15k ohm resistor ---|
- *                                         +--- 10k ohm resistor --- Ground
- *
- *                                         +--- dscReadPin (esp8266: D1, D2, D8)
- *      DSC Green ---- 15k ohm resistor ---|
- *                                         +--- 10k ohm resistor --- Ground
- *
- *  Virtual keypad (optional):
- *      DSC Green ---- NPN collector --\
- *                                      |-- NPN base --- 1k ohm resistor --- dscWritePin (esp8266: D1, D2, D8)
- *            Ground --- NPN emitter --/
- *
- *  Virtual keypad uses an NPN transistor to pull the data line low - most small signal NPN transistors should
- *  be suitable, for example:
- *   -- 2N3904 (EBC)
- *   -- BC547, BC548, BC549
- *
+ * 
+ *  For details on the wiring, please see:
  *  https://github.com/taligentx/dscKeybusInterface
  *
  *
@@ -42,63 +18,89 @@
 
 #include <Homie.h>
 #include <dscKeybusInterface.h>
+#include <time.h>
+#include <ArduinoJson.h>
 
-#define SOFTWARE_VERSION "1.0.1"
+#define SOFTWARE_VERSION "1.1.0 Build 17"
 
 // Configures the Keybus interface with the specified pins
 // dscWritePin is optional, leaving it out disables the virtual keypad.
 #define dscClockPin D1  // esp8266: D1, D2, D8 (GPIO 5, 4, 15)
-#define dscReadPin D2   // esp8266: D1, D2, D8 (GPIO 5, 4, 15)
+#define dscReadPin  D2  // esp8266: D1, D2, D8 (GPIO 5, 4, 15)
 #define dscWritePin D8  // esp8266: D1, D2, D8 (GPIO 5, 4, 15)
 dscKeybusInterface dsc(dscClockPin, dscReadPin, dscWritePin);
 
 enum ArmType { arm_away, arm_stay };
 
-bool homieNodeInputHandler(const String& property, const HomieRange& range, const String& value);
+HomieNode homieNode("alarm", "alarm", "alarm");
+HomieNode *partitions[dscPartitions];
+char partitionNames[dscPartitions][13]; // to store partition names for HomieNode partitions
 
-HomieNode homieNode("alarm", "alarm", homieNodeInputHandler);
 HomieSetting<const char*> dscAccessCode("access-code", "Alarm access code");
 
 unsigned long dscStopTime = 0; // the time (in millis) when the dsc-stop was last requested
 
-void resetDscStatus() {
+byte previousLights[dscPartitions] /*, TODO previousStatus[dscPartitions] */;
+const String bit2str[] = {"OFF", "ON"};
+const char* flagMap[] = { "0", "1" };
+
+const String lightsToJson(const byte lights) {
+  String output;  
+  const size_t capacity = JSON_OBJECT_SIZE(8);
+#if ARDUINOJSON_VERSION_MAJOR >= 6
+  DynamicJsonDocument doc(capacity);
+  UNTESTED
+#else
+  DynamicJsonBuffer jb(capacity); // staticjsonbuffer doesn't seem to work, producing only 6 objects instead of 8
+  JsonObject& doc = jb.createObject();
+#endif
+  doc["ready"]     = bit2str[bitRead(lights, 0)];
+  doc["armed"]     = bit2str[bitRead(lights, 1)];
+  doc["memory"]    = bit2str[bitRead(lights, 2)];
+  doc["bypass"]    = bit2str[bitRead(lights, 3)];
+  doc["trouble"]   = bit2str[bitRead(lights, 4)];
+  doc["program"]   = bit2str[bitRead(lights, 5)];
+  doc["fire"]      = bit2str[bitRead(lights, 6)];
+  doc["backlight"] = bit2str[bitRead(lights, 7)];
+#if ARDUINOJSON_VERSION_MAJOR >= 6
+  serializeJson(doc, output);
+#else
+  doc.printTo(output);
+#endif
+  return output;
+}
+
+void resetStatus() {
+  // we don't want to reset the status of zone changes, which is done by dsc.resetStatus()
   dsc.statusChanged = true;
   dsc.keybusChanged = true;
   dsc.troubleChanged = true;
   dsc.powerChanged = true;
   dsc.batteryChanged = true;
   for (byte partition = 0; partition < dscPartitions; partition++) {
+    dsc.readyChanged[partition] = true;
     dsc.armedChanged[partition] = true;
     dsc.alarmChanged[partition] = true;
-    dsc.exitDelayChanged[partition] = true;
-    dsc.entryDelayChanged[partition] = true;
     dsc.fireChanged[partition] = true;
+    previousLights[partition] = /* TODO previousStatus[i] = */ 0xFF;
   }
-  dsc.openZonesStatusChanged = true;
-  dsc.alarmZonesStatusChanged = true;
-  for (byte zoneGroup = 0; zoneGroup < dscZones; zoneGroup++) {
-    dsc.openZonesChanged[zoneGroup] = dsc.alarmZonesChanged[zoneGroup] = 0xFF;
-  }
-}
-
-void dscEnd() {
-  timer1_detachInterrupt();
-  detachInterrupt(digitalPinToInterrupt(dscClockPin));  
 }
 
 void setupHandler() { 
+  resetStatus();
   dsc.begin();
 }
 
 void loopHandler() {
-  // autmatically restart dsc after 5 minutes if it was stopped
+  // automatically restart dsc after 5 minutes if it was stopped
   if (dscStopTime > 0 && (millis() - dscStopTime) > 5*60*1000) {
     dsc.begin();
     dscStopTime = 0;
     homieNode.setProperty("message").setRetained(false).send("DSC Interface restarted automatically");
   } 
 
-  if (!(dsc.handlePanel() && dsc.statusChanged)) { 
+  dsc.loop();
+  if (!dsc.statusChanged || !dsc.keybusConnected) { 
     return;
   }
   dsc.statusChanged = false;                  
@@ -111,24 +113,24 @@ void loopHandler() {
   }
   
   // Sends the access code when needed by the panel for arming
-  if (dsc.accessCodePrompt && dsc.writeReady) {
+  if (dsc.accessCodePrompt) {
     dsc.accessCodePrompt = false;
     dsc.write(dscAccessCode.get());
   }
 
   if (dsc.troubleChanged) {
     dsc.troubleChanged = false;
-    homieNode.setProperty("trouble").send(dsc.trouble ? "1" : "0");
+    homieNode.setProperty("trouble").send(flagMap[dsc.trouble]);
   }
 
   if (dsc.powerChanged) {
     dsc.powerChanged = false; 
-    homieNode.setProperty("power-trouble").send(dsc.powerTrouble ? "1" : "0");
+    homieNode.setProperty("power-trouble").send(flagMap[dsc.powerTrouble]);
   }
 
   if (dsc.batteryChanged) {
     dsc.batteryChanged = false;
-    homieNode.setProperty("battery-trouble").send(dsc.batteryTrouble ? "1" : "0");
+    homieNode.setProperty("battery-trouble").send(flagMap[dsc.batteryTrouble]);
   }
 
   if (dsc.keypadFireAlarm) {
@@ -146,35 +148,79 @@ void loopHandler() {
     homieNode.setProperty("panic-alarm-keypad").send("1");
   }
 
-  String property;
+  if (dsc.timestampChanged) {
+    dsc.timestampChanged = false;
+    if (!(dsc.year < 0 || dsc.year > 9999 || dsc.month < 1 || dsc.month > 12 || dsc.day < 1 || dsc.day > 31 || dsc.hour < 0 || dsc.hour > 23 || dsc.minute < 0 || dsc.minute > 59)) {
+      char panelTime[17]; // YYYY-MM-DD HH:mm
+      sprintf(panelTime, "%04d-%02d-%02d %02d:%02d", dsc.year, dsc.month, dsc.day, dsc.hour, dsc.minute);
+      homieNode.setProperty("panel-time").setRetained(false).send(panelTime);
+    }
+  }
+
+  String prefix;
   // loop through partitions
   for (byte partition = 0; partition < dscPartitions; partition++) {
-    property = "partition-" + String(partition+1);
+    HomieNode *node = partitions[partition];
+    if (node == nullptr || dsc.status[partition] == 0xC7 || dsc.status[partition] == 0) { // the partition is disabled https://github.com/taligentx/dscKeybusInterface/issues/99
+      continue;
+    }
 
     // Publish exit delay status
     if (dsc.exitDelayChanged[partition]) {
       dsc.exitDelayChanged[partition] = false;
-      homieNode.setProperty(property + "-exit-delay").send(dsc.exitDelay[partition] ? "1" : "0");
+      node->setProperty("exit-delay").send(flagMap[dsc.exitDelay[partition]]);
+    }
+
+    // Publish entry delay status
+    if (dsc.entryDelayChanged[partition]) {
+      dsc.entryDelayChanged[partition] = false;
+      node->setProperty("entry-delay").send(flagMap[dsc.entryDelay[partition]]);
     }
 
     // Publish armed status
     if (dsc.armedChanged[partition]) {
       dsc.armedChanged[partition] = false;
-      homieNode.setProperty(property + "-away").send(dsc.armedAway[partition] ? "1" : "0");
-      homieNode.setProperty(property + "-stay").send(dsc.armedStay[partition] ? "1" : "0");
+      node->setProperty("away").send(flagMap[dsc.armedAway[partition]]);
+      node->setProperty("stay").send(flagMap[dsc.armedStay[partition]]);
     }
 
     // Publish alarm status
     if (dsc.alarmChanged[partition]) {
       dsc.alarmChanged[partition] = false; 
-      homieNode.setProperty(property + "-alarm").send(dsc.alarm[partition] ? "1" : "0");
+      node->setProperty("alarm").send(flagMap[dsc.alarm[partition]]);
     }
 
     // Publish fire alarm status
     if (dsc.fireChanged[partition]) {
       dsc.fireChanged[partition] = false; 
-      homieNode.setProperty(property + "-fire").send(dsc.fire[partition] ? "1" : "0");
+      node->setProperty("fire").send(flagMap[dsc.fire[partition]]);
     }
+
+    // Publish access code
+    if (dsc.accessCodeChanged[partition]) {
+      dsc.accessCodeChanged[partition] = false;
+      String msg;
+      switch (dsc.accessCode[partition]) {
+        case 33:
+        case 34: msg = "duress"; break;
+        case 40: msg = "master"; break;
+        case 41:
+        case 42: msg = "supervisor"; break;
+        default: msg = "user"; 
+      }
+      node->setProperty("access-code").setRetained(false).send(msg);
+    }
+
+    // Publish light status as JSON
+    if (dsc.lights[partition] != previousLights[partition]) {
+      previousLights[partition] = dsc.lights[partition];
+      node->setProperty("lights").send(lightsToJson(dsc.lights[partition]));
+    }
+
+    // if (dsc.status[partition] != previousStatus[partition]) {
+    //   previousStatus[partition] = dsc.status[partition];
+    // TODO
+    // }
   }
 
   // Publish zones 1-64 status
@@ -184,15 +230,15 @@ void loopHandler() {
   //   ...
   //   openZones[7] and openZonesChanged[7]: Bit 0 = Zone 57 ... Bit 7 = Zone 64
   if (dsc.openZonesStatusChanged) {
-    dsc.openZonesStatusChanged = false;                           // Resets the open zones status flag
+    dsc.openZonesStatusChanged = false;    // Reset the open zones status flag
     for (byte zoneGroup = 0; zoneGroup < dscZones; zoneGroup++) {
       for (byte zoneBit = 0; zoneBit < 8; zoneBit++) {
         if (bitRead(dsc.openZonesChanged[zoneGroup], zoneBit)) {  // Checks an individual open zone status flag
-          bitWrite(dsc.openZonesChanged[zoneGroup], zoneBit, 0);  // Resets the individual open zone status flag
-          property = "openzone-" + String(zoneBit + 1 + (zoneGroup * 8));
-          homieNode.setProperty(property).send(bitRead(dsc.openZones[zoneGroup], zoneBit) ? "1" : "0");
+          prefix = "openzone-" + String(zoneBit + 1 + (zoneGroup * 8));
+          homieNode.setProperty(prefix).setRetained(false).send(bitRead(dsc.openZones[zoneGroup], zoneBit) ? "1" : "0");
         }
       }
+      dsc.openZonesChanged[zoneGroup] = 0; // reset the changed flags
     }
   }
 
@@ -207,120 +253,100 @@ void loopHandler() {
     for (byte zoneGroup = 0; zoneGroup < dscZones; zoneGroup++) {
       for (byte zoneBit = 0; zoneBit < 8; zoneBit++) {
         if (bitRead(dsc.alarmZonesChanged[zoneGroup], zoneBit)) {  // Checks an individual alarm zone status flag
-          bitWrite(dsc.alarmZonesChanged[zoneGroup], zoneBit, 0);  // Resets the individual alarm zone status flag
-          property = "alarmzone-" + String(zoneBit + 1 + (zoneGroup * 8));
-          homieNode.setProperty(property).send(bitRead(dsc.alarmZones[zoneGroup], zoneBit) ? "1" : "0");
+          prefix = "alarmzone-" + String(zoneBit + 1 + (zoneGroup * 8));
+          homieNode.setProperty(prefix).send(flagMap[bitRead(dsc.alarmZones[zoneGroup], zoneBit)]);
         }
       }
+      dsc.alarmZonesChanged[zoneGroup] = 0;
     }
   }
 }
 
 bool onKeypad(const HomieRange& range, const String& command) {
-     dsc.write(command.c_str());  
-     return true;
+  // we need to have a static buffer because currently dscKeybusInterface dsc.write() 
+  // works directly with the passed string in inside its interrupt handler, 
+  // so the string given to dsc.write needs to remain in memory
+  static char commandBuffer[64]; 
+  if (command.length() >= sizeof(commandBuffer)/sizeof(commandBuffer[0])) {
+    homieNode.setProperty("message").setRetained(false).send("The keypad data is too long. Max: " + String(sizeof(commandBuffer)-1));
+  } else {
+    unsigned long t = millis();
+    while (!dsc.writeReady) {
+      dsc.loop(); // wait until all pending writes are done
+      if (millis() - t > 10000) { // don't wait forever - max 10s
+        homieNode.setProperty("message").setRetained(false).send("ERROR: Timed out");
+        return true;
+      }
+    } 
+    strcpy(commandBuffer, command.c_str());
+    dsc.write(commandBuffer); 
+  }
+  return true;
 }
 
-// Arm - the partition argument is 0 based
+// Arm - the partition argument is 1 based
 void arm(byte partition, ArmType armType = arm_away) {
-  if (!dsc.armed[partition] && !dsc.exitDelay[partition]) {  // Read the argument sent from the homey flow
-    dsc.writePartition = partition + 1;
-    dsc.write(armType == arm_away ? "w" : "s");  // use the write(char *) version, not the write(char)
+  if (!dsc.armed[partition - 1] && !dsc.exitDelay[partition - 1]) {  // Read the argument sent from the homey flow
+    dsc.writePartition = partition;
+    dsc.write(armType == arm_away ? 'w' : 's'); 
   }
 }
 
-// Disarm - the partition argument is 0 based
+// Disarm - the partition argument is 1 based
 void disarm(byte partition) {
-   if ((dsc.armed[partition] || dsc.exitDelay[partition])) {
-    dsc.writePartition = partition + 1;         // Sets writes to the partition number
+   if ((dsc.armed[partition - 1] || dsc.exitDelay[partition - 1])) {
+    dsc.writePartition = partition;         // Sets writes to the partition number
     dsc.write(dscAccessCode.get());
   }
 }
 
-// returns -1 if no prefix digit is found in str, otherwise 
-// returns the partition prefix (1-8), and remove it from the command string
-// e.g. 1xxx -> return 1, str = "xxx"
-byte extractPrefixDigits(String &str) {
-  byte prefix = -1;
-  unsigned i = 0;
-  while (i < str.length() && isdigit(str[i])) i++;
-  
-  if (i > 0) {
-    String pName = str.substring(0, i);
-    prefix = pName.toInt();
-    str.remove(0, i);
-  }
-  return prefix;
-}
-
-bool homieNodeInputHandler(const String& property, const HomieRange& range, const String& value) {
-  if (!property.startsWith("partition-")) {
-    return false;
-  }
-
-  // property should be in this format: "partition-X-yyy"
-  String prop = property;
-  prop.remove(0, 10); // remove the "partition-" prefix so prop starts with the partition number
-  byte partition = extractPrefixDigits(prop) - 1;
-  // validate partition number
-  if (partition < 0 || partition >= dscPartitions) { // invalid partition number
-    homieNode.setProperty("message").setRetained(false).send("Partition number is out of range (1-" + String(dscPartitions) + ")");
-    return false;
-  }
-
-  prop.remove(0,1); // remove the dash after the digit
-
-  if (prop == "away") {
-    value == "0" || value.equalsIgnoreCase("off") ? disarm(partition) : arm(partition, arm_away);
-    return true;
-  } else if (prop == "stay") {
-    value == "0" || value.equalsIgnoreCase("off") ? disarm(partition) : arm(partition, arm_stay);
-    return true;
-  } 
-  return false;
-}
-
 bool onRefreshStatus(const HomieRange& range, const String& command) {
-  resetDscStatus();
+  resetStatus();
   homieNode.setProperty("refresh-status").setRetained(false).send("OK");
-  return true;
-}
-
-bool onDscActive(const HomieRange& range, const String& command) {
-  if (command == "1") {
-    dsc.begin();
-  } else {
-    dscEnd();
-  }
   return true;
 }
 
 bool onMaintenance(const HomieRange& range, const String& command) {
   if (command == "reboot") {
-    homieNode.setProperty("maintenance").setRetained(false).send("Rebooting...");
+    homieNode.setProperty("message").setRetained(false).send("Rebooting...");
     Homie.reboot();
   } else if (command == "dsc-stop") {
-    dscEnd();
-    homieNode.setProperty("maintenance").setRetained(false).send("DSC Interface stopped");
+    dsc.stop();
     dscStopTime = millis();
-    if (dscStopTime == 0) dscStopTime = 1; // just in case millis() returned 0
+    homieNode.setProperty("message").setRetained(false).send("DSC Interface stopped");
   } else if (command == "dsc-start") {
-    dscStopTime = 0;
-    dsc.begin();
-    homieNode.setProperty("maintenance").setRetained(false).send("DSC Interface started");
+    if (dscStopTime > 0) {
+      dscStopTime = 0;
+      dsc.begin();
+    }
+    homieNode.setProperty("message").setRetained(false).send("DSC Interface started");
   } else {
-    homieNode.setProperty("maintenance").setRetained(false).send("Unknown maintenance command");
+    homieNode.setProperty("message").setRetained(false).send("Unknown maintenance command");
     return false;
+  }
+  return true;
+}
+
+// parse the provided time and set the dsc panel time
+bool onPanelTime(const HomieRange& range, const String& command) {
+  if (dsc.keybusConnected) {
+    struct tm tm; // note tm_year is year since 1900, and tm_mon is month since January. Add offsets accordingly
+    strptime(command.c_str(), "%Y-%m-%d %H:%M", &tm);
+    dsc.setTime(1900 + tm.tm_year, 1 + tm.tm_mon, tm.tm_mday, tm.tm_hour, tm.tm_min, dscAccessCode.get());
+    homieNode.setProperty("message").setRetained(false).send("OK");
+  } else {
+    homieNode.setProperty("message").setRetained(false).send("ERROR: DSC Keybus is not connected");
   }
   return true;
 }
 
 void onHomieEvent(const HomieEvent& event) {
   switch (event.type) {
-    case HomieEventType::MQTT_READY: resetDscStatus(); break;
-    case HomieEventType::OTA_STARTED: dscEnd(); break;
-    case HomieEventType::OTA_SUCCESSFUL:
-    case HomieEventType::OTA_FAILED: dsc.begin(); break;
+    case HomieEventType::MQTT_READY: resetStatus(); break;
+    case HomieEventType::OTA_STARTED: dsc.stop(); break;
+    // case HomieEventType::OTA_SUCCESSFUL:
+    // case HomieEventType::OTA_FAILED: dsc.begin(); break;
+    default: break; // to silence compiler warning
   }
 }
 
@@ -341,15 +367,52 @@ void setup() {
   homieNode.advertise("fire-alarm-keypad");
   homieNode.advertise("aux-alarm-keypad");
   homieNode.advertise("panic-alarm-keypad");
+  homieNode.advertise("panel-time").settable(onPanelTime);
+  homieNode.advertise("debug").settable([](const HomieRange &range, const String &value) {
+    for (byte i = 0; i < dscPartitions; i++) {
+      homieNode.setProperty("debug").setRetained(false).send(String(i) + ": " + String(dsc.status[i], HEX));
+    }
+    return true;
+  });
 
-  for (byte i = 1; i <= dscPartitions; i++) {
-    String pName = "partition-" + String(i) + "-";
-    homieNode.advertise(String(pName + "away").c_str()).settable();
-    homieNode.advertise(String(pName + "stay").c_str()).settable();
-    homieNode.advertise(String(pName + "exit-delay").c_str());
-    homieNode.advertise(String(pName + "alarm").c_str());
-    homieNode.advertise(String(pName + "fire").c_str());
+  dsc.begin();
+  for (int i = 0; i < 3*10; i++) { dsc.loop(); delay(100); }; 
+
+  for (byte i = 0; i < dscPartitions; i++) { 
+    if (dsc.status[i] == 0 || dsc.status[i] == 0xC7) {
+      partitions[i] = nullptr;
+      Serial << "Partition " + String(i+1) + " inactive" << endl;
+      continue;
+    }
+
+    char istr[3];
+    strcpy(partitionNames[i], "partition-");
+    strcat(partitionNames[i], itoa(i+1, istr, 10));
+    HomieNode *node = new HomieNode(partitionNames[i], partitionNames[i], "partition");
+    partitions[i] = node;
+    node->advertise("away").settable([i](const HomieRange &range, const String &value) {
+      if (value == "1" || value.equalsIgnoreCase("on") || value.equalsIgnoreCase("arm")) {
+        arm(i+1, arm_away);
+      } else {
+        disarm(i+1);
+      }
+      return true;
+    });
+    node->advertise("stay").settable([i](const HomieRange &range, const String &value) {
+      if (value == "1" || value.equalsIgnoreCase("on") || value.equalsIgnoreCase("arm")) {
+        arm(i+1, arm_stay);
+      } else {
+        disarm(i+1);
+      }
+      return true;
+    });
+    node->advertise("alarm");
+    node->advertise("exit-delay");
+    node->advertise("entry-delay");
+    node->advertise("fire");
+    node->advertise("access-code");
   }
+  dsc.stop(); // we need to stop it before Homie.setup(), and starts/begins again inside Homie's setupHandler().
 
   Homie.setup();
 }